Re-wrote options_data_collector.py
options_data_collector.py works now.  Re-wrote the different run conditions as methods so it can easily be plugged into other programs. The tickers it tracks is automatically pulled from barcharts.com  with selineum. Keeps a running data-structure of the different tickers along with when we last updated it to be tracked. Gives batches of multiporcessed pools batches of data to maximize the amount of data that TD will return, still have some bugs with that.